requirements.txt
python-consul
redis
PyYaml
ansible
